Implement method to write Category into CSV

+++ src/FileHandler.java
@@ -37,13 +37,26 @@ public class FileHandler {
                 // use string.split to load a string array with the values from
                 // each line of
                 // the file, using a comma as the delimiter
-                String[] attributes = line.split(",");
+                //String[] attributes = line.split(",");
+                String[] attributes = new String[6];
+                int i = 0;
+                for (String piece: line.split(",")) {
+                    if (!piece.equals("") && !piece.equals(null)) {
+                        attributes[i] = piece;
+                    }
+                    i++;
+                }
 
                 InventoryItem inventoryItem = createInventoryItem(attributes);
 
-                // adding inventoryItem into ArrayList
+                // adding inventoryItem into itemMap
                 if (inventoryItem.isValid()) {
                     itemMap.put(inventoryItem.getDescription(), inventoryItem); // only add valid items to inventory
+                }
+                // if Category found
+                // add Category into categoryMap
+                if (inventoryItem.isCategory()) {
+                    App.getInventory().addCategory(new Category(inventoryItem.getCategory()));
                 }
 
                 // read next line before looping
@@ -66,12 +79,24 @@ public class FileHandler {
     private InventoryItem createInventoryItem(String[] metadata) {
         InventoryItem inventoryItem = new InventoryItem(null, null, null, null, null, null);
         try {
-            inventoryItem.description = metadata[0].replaceAll("^\"|\"$", "");
-            inventoryItem.category = metadata[1].replaceAll("^\"|\"$", "");
-            inventoryItem.stock = Integer.parseInt(metadata[2]);
-            inventoryItem.location = metadata[3].replaceAll("^\"|\"$", "");
-            inventoryItem.weight = Integer.parseInt(metadata[4]);
-            inventoryItem.price = Integer.parseInt(metadata[5]);
+            if(metadata[0] != null) {
+                inventoryItem.description = metadata[0].replaceAll("^\"|\"$", "");
+            }
+            if(metadata[1] != null) {
+                inventoryItem.category = metadata[1].replaceAll("^\"|\"$", "");
+            }
+            if(metadata[2] != null) {
+                inventoryItem.stock = Integer.parseInt(metadata[2]);
+            }
+            if(metadata[3] != null) {
+                inventoryItem.location = metadata[3].replaceAll("^\"|\"$", "");
+            }
+            if(metadata[4] != null) {
+                inventoryItem.weight = Integer.parseInt(metadata[4]);
+            }
+            if(metadata[5] != null) {
+                inventoryItem.price = Integer.parseInt(metadata[5]);
+            }
         } catch (Exception e) {
             //e.printStackTrace();
         }
@@ -89,20 +114,20 @@ public class FileHandler {
             String path = inventory.getPath();
             Path file = Paths.get(path);
             String backup = file.getParent().toString() + "/backup.csv";
-            // delete old backup
-            if (Files.exists(Paths.get(backup))) {
-                Files.delete(Paths.get(backup));
-                System.out.println("old backup deleted");
+            if (!path.equals("")) {
+                // delete old backup
+                if (Files.exists(Paths.get(backup))) {
+                    Files.delete(Paths.get(backup));
+                    System.out.println("old backup deleted");
+                }
+                // create backup if old file exists
+                if (Files.exists(file)) {
+                    Files.move(file, Paths.get(backup));
+                    System.out.println("Backup created");
+                } else {
+                    System.out.println("New File created");
+                }
             }
-            // create backup if old file exists
-            if (Files.exists(file)) {
-                Files.move(file, Paths.get(backup));
-                System.out.println("Backup created");
-            }
-            else {
-                System.out.println("New File created");
-            }
-            Files.createFile(file);
 
             // write new file
             // if no path found, write at standard location
@@ -116,10 +141,12 @@ public class FileHandler {
 
             System.out.println("Data saved");
             // delete backup
-            Files.delete(Paths.get(backup));
-            System.out.println("Backup deleted");
+            if (Files.exists(Paths.get(backup))) {
+                Files.delete(Paths.get(backup));
+                System.out.println("Backup deleted");
+            }
         }
-        catch (IOException e) {System.err.println(e.toString());}
+        catch (IOException e) {System.err.println(e.toString()+"test");}
     }
